+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added all filter by dates (https://github.com/AlexcastroDev/flexysearch/issues/4)
+
 ## [0.2.2] - 2022-08-07
 
 ### Fixes